requirements.txt
requests==2.32.5
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.1
